+++ Locale/enUS.lua
@@ -6,6 +6,7 @@ local L = LibStub("AceLocale-3.0"):NewLocale("RCExtraUtilities", "enUS", true)
 if not L then return end
 
 L["Advanced"] = true
+L["Accept Pawn"] = true
 L["Artifact Traits"] = true
 L["Bonus"] = true
 L["Bonus Rolls"] = true
@@ -26,6 +27,7 @@ L["Other"] = true
 L["Pawn specific options"] = true
 L["RCLootCouncil Columns"] = true
 L["Scales"] = true
+L["Score Mode"] = true
 L["Set Pieces"] = true
 L["Sockets"] = true
 L["Spec Icon"] = true
@@ -33,7 +35,10 @@ L["Traits"] = true
 L["Upgrades"] = true
 L["You can't change these settings while the voting frame is showing."] = true
 
+
+L["opt_acceptPawn_desc"] = "Enables candidates to calculate and send their own Pawn scores instead of relying on you to keep your own scales up to date."
 L["opt_pawn_desc"] = "Enables a Pawn column, showing upgrade statistics for the candidate."
+L["opt_pawnMode_desc"] = "Enabling this will display the raw Pawn scores, while disabling will result in upgrade percentage scores."
 L["opt_traits_desc"] = "Enables a column showing the number of artifact traits a candidate has."
 L["opt_upgrades_desc"] = "Enables a column showing x of y valor upgrades a candidate has performed."
 L["opt_sockets_desc"] = "Enables a column showing how many sockets a candidate has on his/her equipped gear."